requirements_dev.txt
-r ./requirements.txt
pytest-cov==4.1.0
requests-mock==1.11.0
black==23.11.0
flake8==6.1.0
isort==5.12.0